directory for parser
