--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "kinolib-backend",
-  "version": "0.0.1",
-  "description": "parser",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/KravaAO/kinolib-backend.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/KravaAO/kinolib-backend/issues"
-  },
-  "homepage": "https://github.com/KravaAO/kinolib-backend#readme",
-  "dependencies": {
-    "axios": "^1.6.8",
-    "cheerio": "^1.0.0-rc.12"
-  }
-}
